refactor(server_media): api-files
